+++ src/NeuralNet/CostFunctions/CrossEntropy.php
@@ -37,11 +37,11 @@ class CrossEntropy implements ClassificationLoss
      *
      * @param \Rubix\Tensor\Tensor $expected
      * @param \Rubix\Tensor\Tensor $output
-     * @return float
+     * @return \Rubix\Tensor\Tensor
      */
-    public function compute(Tensor $expected, Tensor $output) : float
+    public function compute(Tensor $expected, Tensor $output) : Tensor
     {
-        return $expected->negate()->multiply($output->log())->sum()->mean();
+        return $expected->negate()->multiply($output->log());
     }
 
     /**