+++ src/NeuralNet/CostFunctions/HuberLoss.php
@@ -64,12 +64,11 @@ class HuberLoss implements RegressionLoss
      *
      * @param \Rubix\Tensor\Tensor $expected
      * @param \Rubix\Tensor\Tensor $output
-     * @return float
+     * @return \Rubix\Tensor\Tensor
      */
-    public function compute(Tensor $expected, Tensor $output) : float
+    public function compute(Tensor $expected, Tensor $output) : Tensor
     {
-        return $expected->subtract($output)
-            ->map([$this, '_compute'])->sum()->mean();
+        return $expected->subtract($output)->map([$this, '_compute']);
     }
 
     /**