+++ src/NeuralNet/Layers/Binary.php
@@ -289,7 +289,7 @@ class Binary implements Output
 
         $gradient = new Deferred([$this, 'gradient'], [$w, $dA]);
 
-        $loss = $this->costFn->compute($expected, $this->computed);
+        $loss = $this->costFn->compute($expected, $this->computed)->sum()->mean();
 
         unset($this->input, $this->z, $this->computed);