+++ tests/NeuralNet/CostFunctions/RelativeEntropyTest.php
@@ -32,16 +32,22 @@ class RelativeEntropyTest extends TestCase
 
     public function test_compute()
     {
-        $cost = $this->costFn->compute($this->expected, $this->activation);
+        $loss = $this->costFn->compute($this->expected, $this->activation)->asArray();
 
-        $this->assertEquals(0.04663867866572951, $cost);
+        $expected = [
+            [0.010050335853501506],
+            [-1.6811242831518263E-7],
+            [-1.8064005800013633E-7],
+            [0.22314355131420976],
+            [-1.4508657738524219E-7],
+        ];
+
+        $this->assertEquals($expected, $loss);
     }
 
     public function test_differentiate()
     {
-        $derivative = $this->costFn
-            ->differentiate($this->expected, $this->activation)
-            ->asArray();
+        $gradient = $this->costFn->differentiate($this->expected, $this->activation)->asArray();
 
         $expected = [
             [-0.01010101010101011],
@@ -51,6 +57,6 @@ class RelativeEntropyTest extends TestCase
             [0.9999994999999999],
         ];
 
-        $this->assertEquals($expected, $derivative);
+        $this->assertEquals($expected, $gradient);
     }
 }