+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Unreleased
+    - Added Regression and Classification Loss interfaces
     - LODA now defaults to auto detect bin count
 
 - 0.0.12-beta